+++ positioning/src/main/java/com/getwemap/example/positioning/Constants.kt
@@ -1,8 +1,0 @@
-package com.getwemap.example.positioning
-
-object Constants {
-
-    val vpsEndpoint: String get() {
-        throw Exception("Specify vpsEndpoint and remove fatalError")
-    }
-}